[fix] accountのGETリクエストのときのハンドウルを別のファイルにする
