+++ app/handler/accounts/create.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"yatter-backend-go/app/domain/object"
-	"github.com/go-chi/chi/v5"
 )
 
 // Request body for `POST /v1/accounts`
@@ -33,34 +32,6 @@ func (h *handler) Create(w http.ResponseWriter, r *http.Request) {
 
 	if err := h.ar.CreateUser(r.Context(), account); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	if err := json.NewEncoder(w).Encode(account); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-// Handle request for `GET /v1/accounts/{username}`
-func (h *handler) GetUser(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	username := chi.URLParam(r, "username")
-	if username == "" {
-		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
-		return
-	}
-
-	account, err := h.ar.FindByUsername(ctx, username)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if account == nil {
-		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
