+++ app/handler/accounts/get.go
@@ -1,0 +1,36 @@
+package accounts
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/go-chi/chi/v5"
+)
+
+// Handle request for `GET /v1/accounts/{username}`
+func (h *handler) GetUser(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+
+	username := chi.URLParam(r, "username")
+	if username == "" {
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+
+	account, err := h.ar.FindByUsername(ctx, username)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	if account == nil {
+		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	if err := json.NewEncoder(w).Encode(account); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
